refactor(cli): remove obsolete node timeout overrides
